fix: moving return to command button to top of investigate
