+++ src/Investigate/Components/Investigate.tsx
@@ -4,13 +4,31 @@ import Mnemonics from "./Mnemonics/Mnemonics";
 import "./Investigate.css";
 
 import { useAppContext, ContextType } from "provider/useAppContext";
+import { RuxButton } from "@astrouxds/react";
 
 const Investigate = () => {
-  const { showInvestigate, selectedAssemblyDevice }: ContextType =
-    useAppContext();
+  const {
+    showInvestigate,
+    selectedAssemblyDevice,
+    toggleInvestigate,
+    resetSelected,
+  }: ContextType = useAppContext();
+
+  const handleReturnToCommand = () => {
+    toggleInvestigate();
+    resetSelected();
+  };
 
   return (
     <div className="investigate-background" data-active={showInvestigate}>
+      <RuxButton
+        className="investigate-breadcrumb"
+        borderless
+        icon="keyboard-arrow-left"
+        onClick={() => handleReturnToCommand()}
+      >
+        Return to Command
+      </RuxButton>
       <SubsystemsTree />
       <Assembly />
       <Mnemonics title={selectedAssemblyDevice.name} />
